fix(openvino): Resolve runtime errors and refactor model structure
This commit addresses several critical runtime errors that occurred when using the OpenVINO model source and refactors the location of the model handler scripts for better portability..

MODEL REFACTORING:
- The model handler scripts (`BaseWhisper.py`, `FasterWhisper.py`, `OpenVINOWhisper.py`) have been moved from the `modules/models/` subdirectory to the main `modules/` directory.
- **Reason:** The previous nested structure caused issues with pathing in the portable (frozen) executable. This change flattens the structure to ensure reliability.
- Updated all relevant import paths in `transcribe_audio.py` to match the new location.

OPEN-VINO FIXES:
- **Language Name Bug:** Resolved a `TypeError` caused by an incorrect call to `pycountry.languages.get()`. The code now correctly uses the `name=` keyword to look up full language names (e.g., "English") and convert them to the required two-letter ISO code.
- **Device Mismatch Bug:** Fixed a `TypeError: can't convert cuda:0 device type tensor to numpy`. The `transformers.pipeline` was creating tensors on the GPU even when `--device cpu` was specified. The pipeline is now explicitly passed the CPU device, ensuring all operations remain on the host memory as intended.
- **Device Case-Sensitivity Bug:** Corrected a `RuntimeError` where the uppercase "CPU" (required by OpenVINO) was incompatible with the `transformers` library (which expects lowercase "cpu"). The device string is now converted to lowercase just before being passed to the pipeline.
