+++ blacklist.txt
@@ -15,3 +15,4 @@ subscribe to my channel
 Thank you for watching.
 Thank you
 Hi guys, welcome back to my channel
+Thanks for watching!
