+++ modules/BaseWhisper.py
