+++ transcribe_audio.py
@@ -40,9 +40,9 @@ if sys.platform.startswith('win'):
 from modules.audio_handlers import record_callback, handle_mic_calibration
 from modules.device_manager import get_microphone_source, list_microphones, setup_device
 from modules.file_handlers import load_blacklist, setup_temp_directory, clean_temp_directory, save_transcript, handle_error, cleanup_temp_cookie_file
-from modules.models.BaseWhisper import BaseWhisperModel
-from modules.models.FasterWhisper import FasterWhisperModel
-from modules.models.OpenVINOWhisper import OpenVINOWhisperModel
+from modules.BaseWhisper import BaseWhisperModel
+from modules.FasterWhisper import FasterWhisperModel
+from modules.OpenVINOWhisper import OpenVINOWhisperModel
 from modules.transcription_core import TranscriptionCore
 from modules.stream_handler import handle_stream_setup
 from modules.stream_transcription_module import stop_transcription
